Analyst has an Add Resource modal [#160921407]

+++ scripts/setup_acceptance.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+if [[ "${@}" == *"--help"* ]]; then
+    echo "./setup_acceptance [--build] [--reseed]"
+    exit 0
+fi
+
 BASE_DIR="$(dirname $( cd "$(dirname "$0")" ; pwd -P ))"
 PROGRAM_NAME="BlackPearlAcceptance"
 
@@ -9,10 +14,11 @@ echo "Stopping Docker Container"
 docker stop ${PROGRAM_NAME} || true
 #docker rm ${PROGRAM_NAME} || true
 
-${BASE_DIR}/scripts/build_jar.sh --no-replace
-
-rm ${BASE_DIR}/acceptance/blackpearl-[0-9\.]*-SNAPSHOT.jar
-cp ${BASE_DIR}/target/blackpearl-[0-9\.]*-SNAPSHOT.jar ${BASE_DIR}/acceptance/
+if [[ "${@}" == *"--build"* ]]; then
+    ${BASE_DIR}/scripts/build_jar.sh --no-replace
+    rm ${BASE_DIR}/acceptance/blackpearl-[0-9\.]*-SNAPSHOT.jar
+    cp ${BASE_DIR}/target/blackpearl-[0-9\.]*-SNAPSHOT.jar ${BASE_DIR}/acceptance/
+fi
 
 if [ $(docker container list -a | grep -i ${PROGRAM_NAME} | wc -l) -eq 0 ]; then
   docker run --name ${PROGRAM_NAME} -v ${BASE_DIR}:/app -p 7070:7070 -itd dgs1sdt/blackpearl
@@ -22,12 +28,6 @@ fi
 
 docker exec -d ${PROGRAM_NAME} /bin/bash -c "java -jar /app/acceptance/blackpearl-[0-9\.]*-SNAPSHOT.jar --server.port=7070"
 
-printf "Would you like to seed the database [Y/N]:"
-read seedDatabase
-
-if [ "${seedDatabase}" == "Y" ] || [ "${seedDatabase}" == "y" ]; then
+if [[ "${@}" == "--reseed" ]]; then
   docker exec -d ${PROGRAM_NAME} /bin/bash -c "/app/scripts/seed_db/seed_db.sh"
-  echo "Database Reseeded"
-else
-  echo "Database Not Reseeded"
 fi
